Normalize decoded URLs before displaying them
This change makes it impossible to hide the URL in the UI by prefixing it with a bunch of newlines.

Detecting `javascript:` URLs and all their variations such as `JaVaScRiPt:` also becomes much easier. This makes it possible to programmatically detect XSS attempts and to handle them accordingly (e.g. show a warning or error message).