Add option to serve the app as dev
By serving the raw application files, there is no longer a need for
bundling during development. In this case, a refresh of the page will
always load the last version of the files.

This also makes use of some new web features, which are (at the time of
writing) in little to no browsers. (Most notably, it uses ES modules in
workers, albeit not in SW, see:
https://bugs.chromium.org/p/chromium/issues/detail?id=824647)

You can test this out by running `npm run start-dev` and open
`localhost:8080` in your web browser. You can simply change any file you
like and hit refresh to see the updates.