Improve feature detection and reuse BarcodeDetector
This patch includes improved feature detection support to decide whether
or not the polyfill is necessary. It first checks if the BarcodeDetector
supports QR codes and if so creates a new instance configured to only
detect this format. This instance is reused between frames to avoid
repeating the initialization steps.

Usage of the BarcodeDetector object is also moved to the main thread.
While the polyfill is blocking the native implementation is asynchronous
and sending the image data to a worker first is wasted effort.